#520 Move OffsetCommitRequest to the model package.
